+++ www/wp-content/themes/hittheroad-2021/woocommerce/myaccount/form-lost-password.php
@@ -12,7 +12,7 @@
  *
  * @see https://docs.woocommerce.com/document/template-structure/
  * @package WooCommerce\Templates
- * @version 7.0.1
+ * @version 9.0.0
  */
 
 defined('ABSPATH') || exit; ?>
@@ -27,7 +27,7 @@ do_action('woocommerce_before_lost_password_form');
 		<p><?= apply_filters('woocommerce_lost_password_message', esc_html__('Lost your password? Please enter your username or email address. You will receive a link to create a new password via email.', 'woocommerce')); ?></p><?php // @codingStandardsIgnoreLine ?>
 
 		<div class="form-floating mb-3">
-			<input class="form-control" type="text" name="user_login" id="user_login" autocomplete="username" placeholder="Login">
+			<input class="form-control" type="text" name="user_login" id="user_login" autocomplete="username" placeholder="Login" aria-required="true">
 			<label for="user_login"><?php esc_html_e('Username or email', 'woocommerce'); ?></label>
 		</div>