+++ www/wp-content/themes/hittheroad-2021/woocommerce/myaccount/navigation.php
@@ -12,7 +12,7 @@
  *
  * @see     https://docs.woocommerce.com/document/template-structure/
  * @package WooCommerce\Templates
- * @version 2.6.0
+ * @version 9.0.0
  */
 
 if (!defined('ABSPATH')) {
@@ -22,7 +22,7 @@ if (!defined('ABSPATH')) {
 do_action('woocommerce_before_account_navigation');
 ?>
 
-<nav class="my-account-nav">
+<nav class="my-account-nav" aria-label="<?php esc_html_e('Account pages', 'woocommerce'); ?>">
 	<ul>
 		<?php foreach (wc_get_account_menu_items() as $endpoint => $label) : ?>
 			<li class="<?= wc_get_account_menu_item_classes($endpoint); ?>">